+++ src/apps/depth2x/converter_max_curve.h.inl
@@ -1,6 +1,6 @@
 template <typename Intrinsic>
 bool max_curve_converter<Intrinsic>::process_file(
-    const math::image<double>& depth_image, int idx) const noexcept {
+    const math::image<float>& depth_image, int idx) const noexcept {
     Expects(!this->_files.output.empty());
     using namespace conversion;
